Implemented syscall write.
This implementation expects a complete message string to be passed as an
argument. The string is then copied over to an internal buffer and
passed to the serial driver. Subsequent calls will fail until the driver
releases the buffer. This works poorly with newlib nano, since it
typically makes multiple single character calls to write.